+++ authpass/lib/env/_base.freezed.dart
@@ -1,0 +1,316 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+
+part of '_base.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+class _$EnvSecretsTearOff {
+  const _$EnvSecretsTearOff();
+
+// ignore: unused_element
+  _EnvSecrets call(
+      {@required @nullable String analyticsAmplitudeApiKey,
+      @required @nullable String analyticsGoogleAnalyticsId,
+      @required @nullable String googleClientId,
+      @required @nullable String googleClientSecret,
+      @required @nullable String dropboxKey,
+      @required @nullable String dropboxSecret,
+      @required @nullable String microsoftClientId,
+      @required @nullable String microsoftClientSecret}) {
+    return _EnvSecrets(
+      analyticsAmplitudeApiKey: analyticsAmplitudeApiKey,
+      analyticsGoogleAnalyticsId: analyticsGoogleAnalyticsId,
+      googleClientId: googleClientId,
+      googleClientSecret: googleClientSecret,
+      dropboxKey: dropboxKey,
+      dropboxSecret: dropboxSecret,
+      microsoftClientId: microsoftClientId,
+      microsoftClientSecret: microsoftClientSecret,
+    );
+  }
+}
+
+// ignore: unused_element
+const $EnvSecrets = _$EnvSecretsTearOff();
+
+mixin _$EnvSecrets {
+  @nullable
+  String get analyticsAmplitudeApiKey;
+  @nullable
+  String get analyticsGoogleAnalyticsId;
+  @nullable
+  String get googleClientId;
+  @nullable
+  String get googleClientSecret;
+  @nullable
+  String get dropboxKey;
+  @nullable
+  String get dropboxSecret;
+  @nullable
+  String get microsoftClientId;
+  @nullable
+  String get microsoftClientSecret;
+
+  $EnvSecretsCopyWith<EnvSecrets> get copyWith;
+}
+
+abstract class $EnvSecretsCopyWith<$Res> {
+  factory $EnvSecretsCopyWith(
+          EnvSecrets value, $Res Function(EnvSecrets) then) =
+      _$EnvSecretsCopyWithImpl<$Res>;
+  $Res call(
+      {@nullable String analyticsAmplitudeApiKey,
+      @nullable String analyticsGoogleAnalyticsId,
+      @nullable String googleClientId,
+      @nullable String googleClientSecret,
+      @nullable String dropboxKey,
+      @nullable String dropboxSecret,
+      @nullable String microsoftClientId,
+      @nullable String microsoftClientSecret});
+}
+
+class _$EnvSecretsCopyWithImpl<$Res> implements $EnvSecretsCopyWith<$Res> {
+  _$EnvSecretsCopyWithImpl(this._value, this._then);
+
+  final EnvSecrets _value;
+  // ignore: unused_field
+  final $Res Function(EnvSecrets) _then;
+
+  @override
+  $Res call({
+    Object analyticsAmplitudeApiKey = freezed,
+    Object analyticsGoogleAnalyticsId = freezed,
+    Object googleClientId = freezed,
+    Object googleClientSecret = freezed,
+    Object dropboxKey = freezed,
+    Object dropboxSecret = freezed,
+    Object microsoftClientId = freezed,
+    Object microsoftClientSecret = freezed,
+  }) {
+    return _then(_value.copyWith(
+      analyticsAmplitudeApiKey: analyticsAmplitudeApiKey == freezed
+          ? _value.analyticsAmplitudeApiKey
+          : analyticsAmplitudeApiKey as String,
+      analyticsGoogleAnalyticsId: analyticsGoogleAnalyticsId == freezed
+          ? _value.analyticsGoogleAnalyticsId
+          : analyticsGoogleAnalyticsId as String,
+      googleClientId: googleClientId == freezed
+          ? _value.googleClientId
+          : googleClientId as String,
+      googleClientSecret: googleClientSecret == freezed
+          ? _value.googleClientSecret
+          : googleClientSecret as String,
+      dropboxKey:
+          dropboxKey == freezed ? _value.dropboxKey : dropboxKey as String,
+      dropboxSecret: dropboxSecret == freezed
+          ? _value.dropboxSecret
+          : dropboxSecret as String,
+      microsoftClientId: microsoftClientId == freezed
+          ? _value.microsoftClientId
+          : microsoftClientId as String,
+      microsoftClientSecret: microsoftClientSecret == freezed
+          ? _value.microsoftClientSecret
+          : microsoftClientSecret as String,
+    ));
+  }
+}
+
+abstract class _$EnvSecretsCopyWith<$Res> implements $EnvSecretsCopyWith<$Res> {
+  factory _$EnvSecretsCopyWith(
+          _EnvSecrets value, $Res Function(_EnvSecrets) then) =
+      __$EnvSecretsCopyWithImpl<$Res>;
+  @override
+  $Res call(
+      {@nullable String analyticsAmplitudeApiKey,
+      @nullable String analyticsGoogleAnalyticsId,
+      @nullable String googleClientId,
+      @nullable String googleClientSecret,
+      @nullable String dropboxKey,
+      @nullable String dropboxSecret,
+      @nullable String microsoftClientId,
+      @nullable String microsoftClientSecret});
+}
+
+class __$EnvSecretsCopyWithImpl<$Res> extends _$EnvSecretsCopyWithImpl<$Res>
+    implements _$EnvSecretsCopyWith<$Res> {
+  __$EnvSecretsCopyWithImpl(
+      _EnvSecrets _value, $Res Function(_EnvSecrets) _then)
+      : super(_value, (v) => _then(v as _EnvSecrets));
+
+  @override
+  _EnvSecrets get _value => super._value as _EnvSecrets;
+
+  @override
+  $Res call({
+    Object analyticsAmplitudeApiKey = freezed,
+    Object analyticsGoogleAnalyticsId = freezed,
+    Object googleClientId = freezed,
+    Object googleClientSecret = freezed,
+    Object dropboxKey = freezed,
+    Object dropboxSecret = freezed,
+    Object microsoftClientId = freezed,
+    Object microsoftClientSecret = freezed,
+  }) {
+    return _then(_EnvSecrets(
+      analyticsAmplitudeApiKey: analyticsAmplitudeApiKey == freezed
+          ? _value.analyticsAmplitudeApiKey
+          : analyticsAmplitudeApiKey as String,
+      analyticsGoogleAnalyticsId: analyticsGoogleAnalyticsId == freezed
+          ? _value.analyticsGoogleAnalyticsId
+          : analyticsGoogleAnalyticsId as String,
+      googleClientId: googleClientId == freezed
+          ? _value.googleClientId
+          : googleClientId as String,
+      googleClientSecret: googleClientSecret == freezed
+          ? _value.googleClientSecret
+          : googleClientSecret as String,
+      dropboxKey:
+          dropboxKey == freezed ? _value.dropboxKey : dropboxKey as String,
+      dropboxSecret: dropboxSecret == freezed
+          ? _value.dropboxSecret
+          : dropboxSecret as String,
+      microsoftClientId: microsoftClientId == freezed
+          ? _value.microsoftClientId
+          : microsoftClientId as String,
+      microsoftClientSecret: microsoftClientSecret == freezed
+          ? _value.microsoftClientSecret
+          : microsoftClientSecret as String,
+    ));
+  }
+}
+
+class _$_EnvSecrets implements _EnvSecrets {
+  const _$_EnvSecrets(
+      {@required @nullable this.analyticsAmplitudeApiKey,
+      @required @nullable this.analyticsGoogleAnalyticsId,
+      @required @nullable this.googleClientId,
+      @required @nullable this.googleClientSecret,
+      @required @nullable this.dropboxKey,
+      @required @nullable this.dropboxSecret,
+      @required @nullable this.microsoftClientId,
+      @required @nullable this.microsoftClientSecret});
+
+  @override
+  @nullable
+  final String analyticsAmplitudeApiKey;
+  @override
+  @nullable
+  final String analyticsGoogleAnalyticsId;
+  @override
+  @nullable
+  final String googleClientId;
+  @override
+  @nullable
+  final String googleClientSecret;
+  @override
+  @nullable
+  final String dropboxKey;
+  @override
+  @nullable
+  final String dropboxSecret;
+  @override
+  @nullable
+  final String microsoftClientId;
+  @override
+  @nullable
+  final String microsoftClientSecret;
+
+  @override
+  String toString() {
+    return 'EnvSecrets(analyticsAmplitudeApiKey: $analyticsAmplitudeApiKey, analyticsGoogleAnalyticsId: $analyticsGoogleAnalyticsId, googleClientId: $googleClientId, googleClientSecret: $googleClientSecret, dropboxKey: $dropboxKey, dropboxSecret: $dropboxSecret, microsoftClientId: $microsoftClientId, microsoftClientSecret: $microsoftClientSecret)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other is _EnvSecrets &&
+            (identical(
+                    other.analyticsAmplitudeApiKey, analyticsAmplitudeApiKey) ||
+                const DeepCollectionEquality().equals(
+                    other.analyticsAmplitudeApiKey,
+                    analyticsAmplitudeApiKey)) &&
+            (identical(other.analyticsGoogleAnalyticsId,
+                    analyticsGoogleAnalyticsId) ||
+                const DeepCollectionEquality().equals(
+                    other.analyticsGoogleAnalyticsId,
+                    analyticsGoogleAnalyticsId)) &&
+            (identical(other.googleClientId, googleClientId) ||
+                const DeepCollectionEquality()
+                    .equals(other.googleClientId, googleClientId)) &&
+            (identical(other.googleClientSecret, googleClientSecret) ||
+                const DeepCollectionEquality()
+                    .equals(other.googleClientSecret, googleClientSecret)) &&
+            (identical(other.dropboxKey, dropboxKey) ||
+                const DeepCollectionEquality()
+                    .equals(other.dropboxKey, dropboxKey)) &&
+            (identical(other.dropboxSecret, dropboxSecret) ||
+                const DeepCollectionEquality()
+                    .equals(other.dropboxSecret, dropboxSecret)) &&
+            (identical(other.microsoftClientId, microsoftClientId) ||
+                const DeepCollectionEquality()
+                    .equals(other.microsoftClientId, microsoftClientId)) &&
+            (identical(other.microsoftClientSecret, microsoftClientSecret) ||
+                const DeepCollectionEquality().equals(
+                    other.microsoftClientSecret, microsoftClientSecret)));
+  }
+
+  @override
+  int get hashCode =>
+      runtimeType.hashCode ^
+      const DeepCollectionEquality().hash(analyticsAmplitudeApiKey) ^
+      const DeepCollectionEquality().hash(analyticsGoogleAnalyticsId) ^
+      const DeepCollectionEquality().hash(googleClientId) ^
+      const DeepCollectionEquality().hash(googleClientSecret) ^
+      const DeepCollectionEquality().hash(dropboxKey) ^
+      const DeepCollectionEquality().hash(dropboxSecret) ^
+      const DeepCollectionEquality().hash(microsoftClientId) ^
+      const DeepCollectionEquality().hash(microsoftClientSecret);
+
+  @override
+  _$EnvSecretsCopyWith<_EnvSecrets> get copyWith =>
+      __$EnvSecretsCopyWithImpl<_EnvSecrets>(this, _$identity);
+}
+
+abstract class _EnvSecrets implements EnvSecrets {
+  const factory _EnvSecrets(
+      {@required @nullable String analyticsAmplitudeApiKey,
+      @required @nullable String analyticsGoogleAnalyticsId,
+      @required @nullable String googleClientId,
+      @required @nullable String googleClientSecret,
+      @required @nullable String dropboxKey,
+      @required @nullable String dropboxSecret,
+      @required @nullable String microsoftClientId,
+      @required @nullable String microsoftClientSecret}) = _$_EnvSecrets;
+
+  @override
+  @nullable
+  String get analyticsAmplitudeApiKey;
+  @override
+  @nullable
+  String get analyticsGoogleAnalyticsId;
+  @override
+  @nullable
+  String get googleClientId;
+  @override
+  @nullable
+  String get googleClientSecret;
+  @override
+  @nullable
+  String get dropboxKey;
+  @override
+  @nullable
+  String get dropboxSecret;
+  @override
+  @nullable
+  String get microsoftClientId;
+  @override
+  @nullable
+  String get microsoftClientSecret;
+  @override
+  _$EnvSecretsCopyWith<_EnvSecrets> get copyWith;
+}